lowering: register core op lowerings
