+++ src/onnx2c/lowering/common.py
@@ -1,0 +1,90 @@
+from __future__ import annotations
+
+from collections.abc import Sequence
+
+from ..errors import ShapeInferenceError, UnsupportedOpError
+from ..ir.model import Graph, Node
+
+
+def _ensure_supported_dtype(dtype: str) -> str:
+    if dtype not in {
+        "float",
+        "double",
+        "bool",
+        "int64",
+        "int32",
+        "int16",
+        "int8",
+        "uint64",
+        "uint32",
+        "uint16",
+        "uint8",
+    }:
+        raise UnsupportedOpError(f"Unsupported dtype {dtype}")
+    return dtype
+
+
+def value_dtype(graph: Graph, name: str, node: Node) -> str:
+    try:
+        value = graph.find_value(name)
+    except KeyError as exc:
+        raise ShapeInferenceError(
+            f"Missing dtype for value '{name}' in op {node.op_type}. "
+            "Hint: run ONNX shape inference or export with static shapes."
+        ) from exc
+    return _ensure_supported_dtype(value.type.dtype)
+
+
+def value_shape(graph: Graph, name: str, node: Node) -> tuple[int, ...]:
+    try:
+        return graph.find_value(name).type.shape
+    except KeyError as exc:
+        raise ShapeInferenceError(
+            f"Missing shape for value '{name}' in op {node.op_type}. "
+            "Hint: run ONNX shape inference or export with static shapes."
+        ) from exc
+
+
+def node_dtype(graph: Graph, node: Node, *names: str) -> str:
+    filtered = [name for name in names if name]
+    if not filtered:
+        raise UnsupportedOpError(
+            f"{node.op_type} expects at least one typed input or output"
+        )
+    dtypes = {value_dtype(graph, name, node) for name in filtered}
+    if len(dtypes) != 1:
+        raise UnsupportedOpError(
+            f"{node.op_type} expects matching dtypes, got {', '.join(sorted(dtypes))}"
+        )
+    return next(iter(dtypes))
+
+
+def shape_product(shape: tuple[int, ...]) -> int:
+    if not shape:
+        return 1
+    product = 1
+    for dim in shape:
+        if dim <= 0:
+            raise ShapeInferenceError("Dynamic or zero dims are not supported")
+        product *= dim
+    return product
+
+
+def normalize_axis(axis: int, shape: tuple[int, ...], node: Node) -> int:
+    if not shape:
+        raise ShapeInferenceError(f"{node.op_type} does not support scalar inputs")
+    rank = len(shape)
+    if axis < 0:
+        axis += rank
+    if axis < 0 or axis >= rank:
+        raise ShapeInferenceError(
+            f"{node.op_type} axis {axis} is out of range for rank {rank}"
+        )
+    return axis
+
+
+def optional_name(names: Sequence[str], index: int) -> str | None:
+    if index >= len(names):
+        return None
+    name = names[index]
+    return name or None
